+++ crates/ruma-common/CHANGELOG.md
@@ -8,6 +8,7 @@ Bug fixes:
   and `events::secret` modules
 * Fix deserialization of `RoomMessageEventContent` and `RoomEncryptedEventContent` when there
   is no relation
+* Fix deserialization of `StateUnsigned` when the `prev_content` is redacted
 
 Breaking changes: